Applied updates and added column, record and table OSS-Fuzz targets

+++ pyesedb/pyesedb_datetime.h
@@ -37,7 +37,7 @@ PyObject *pyesedb_datetime_new_from_time_elements(
            uint8_t hours,
            uint8_t minutes,
            uint8_t seconds,
-           uint8_t micro_seconds );
+           uint32_t micro_seconds );
 
 PyObject *pyesedb_datetime_new_from_fat_date_time(
            uint32_t fat_date_time );
